update notes with PATCH request

+++ SimpleNotesApp/Home.swift
@@ -8,10 +8,14 @@
 import SwiftUI
 
 struct Home: View {
-    @State var notes = [Note]()
-    @State var showAdd = false
-    @State var showAlert = false
-    @State var deleteItem: Note?
+    @State private var notes = [Note]()
+    @State private var showAdd = false
+    @State private var showAlert = false
+    @State private var deleteItem: Note?
+    @State private var updateNote = ""
+    @State private var updateNoteId = ""
+    
+    @State private var isEditMode: EditMode = .inactive
     
     var alert: Alert {
         Alert(title: Text("Delete"), message: Text("Are you sure you want to delete this note?"), primaryButton: .destructive(Text("Delete"), action: deleteNote), secondaryButton: .cancel())
@@ -20,28 +24,59 @@ struct Home: View {
     var body: some View {
         NavigationView {
             List(self.notes) { note in
-                Text(note.note)
-                    .padding()
-                    .onLongPressGesture {
-                        self.showAlert.toggle()
-                        deleteItem = note
+                if (self.isEditMode == .inactive) {
+                    Text(note.note)
+                        .padding()
+                        .onLongPressGesture {
+                            self.showAlert.toggle()
+                            deleteItem = note
+                        }
+                } else {
+                    HStack {
+                        Image(systemName: "pencil.circle.fill")
+                            .foregroundColor(.yellow)
+                        
+                        Text(note.note)
+                            .padding()
                     }
+                    .onTapGesture {
+                        self.updateNote = note.note
+                        self.updateNoteId = note.id
+                        self.showAdd.toggle()
+                    }
+                }
             }
             .alert(isPresented: $showAlert, content: {
                 alert
-            })
-            .sheet(isPresented: $showAdd, onDismiss: fetchNotes, content: {
-                AddNoteView()
             })
             .onAppear(perform: {
                 fetchNotes()
             })
             .navigationTitle("Notes")
-            .navigationBarItems(trailing: Button(action: {
+            .navigationBarItems(leading: Button(action: {
+                if (self.isEditMode == .inactive) {
+                    self.isEditMode = .active
+                } else {
+                    self.isEditMode = .inactive
+                }
+            }, label: {
+                if (self.isEditMode == .inactive) {
+                    Text("Edit")
+                } else {
+                    Text("Done")
+                }
+            }), trailing: Button(action: {
                 self.showAdd.toggle()
             }, label: {
                 Text("Add")
             }))
+            .sheet(isPresented: $showAdd, onDismiss: fetchNotes, content: {
+                if (self.isEditMode == .inactive) {
+                    AddNoteView()
+                } else {
+                    UpdateNoteView(note: $updateNote, noteId: $updateNoteId)
+                }
+            })
         }
     }
     
@@ -63,6 +98,10 @@ struct Home: View {
         }
         
         task.resume()
+        
+        if (self.isEditMode == .active) {
+            self.isEditMode = .inactive
+        }
     }
     
     func deleteNote() {
